refactored packages for stats stack
